+++ changelog.md
@@ -1,6 +1,5 @@
 # 1.0.2
-- Literally add one line of code that should only be called if something goes wrong.
-Please add a sub-update letter to the version system. /hj
+- Add makeshift opacity approximation for blending as a setting since doing additive blending on a transparent background is impossible.
 
 # 1.0.1
 - Make a few changes to the code (shouldn't affect the in-game outcome)