+++ changelog.md
@@ -1,5 +1,6 @@
 # 1.0.2
 - Add makeshift opacity approximation for blending as a setting since doing additive blending on a transparent background is impossible.
+- Fix crash when exporting with no selected objects
 
 # 1.0.1
 - Make a few changes to the code (shouldn't affect the in-game outcome)